🔒Security : 블로그 글 인가 관련 수정

+++ src/main/java/com/portalSite/blog/service/BlogPostService.java
@@ -46,6 +46,10 @@ public class BlogPostService {
         Member blogMember = memberRepository.findById(memberId)
             .orElseThrow(() -> new RuntimeException("존재하지 않는 사용자입니다."));
 
+        if (blog.getMember().equals(blogMember)) {
+            throw new RuntimeException("다른 사람의 블로그입니다.");
+        }
+
         BlogPost blogPost = BlogPost.of(blog, blogBoard, blogMember, request.getTitle(),
             request.getDescription());
 
@@ -74,10 +78,13 @@ public class BlogPostService {
         return BlogPostResponse.from(blogPost);
     }
 
-    public BlogPostResponse updateBlogPost(UpdateBlogPostRequest request, Long blogPostId) {
+    public BlogPostResponse updateBlogPost(UpdateBlogPostRequest request, Long blogPostId, Long memberId) {
         BlogPost blogPost = blogPostRepository.findById(blogPostId)
             .orElseThrow(() -> new RuntimeException("존재하지 않는 게시글입니다."));
 
+        if (blogPost.getMember().getId().equals(memberId)) {
+            throw new RuntimeException("다른 사람의 블로그입니다.");
+        }
         BlogBoard blogBoard = Optional.ofNullable(request.getBlogBoardId())
             .flatMap(blogBoardRepository::findById)
             .orElse(null);
@@ -88,9 +95,14 @@ public class BlogPostService {
         return BlogPostResponse.from(blogPostRepository.save(connectedPost));
     }
 
-    public void deleteBlogPost(Long blogPostId) {
+    public void deleteBlogPost(Long blogPostId, Long memberId) {
+
         BlogPost blogPost = blogPostRepository.findById(blogPostId)
             .orElseThrow(() -> new RuntimeException("존재하지 않는 게시물입니다."));
+
+        if (blogPost.getMember().getId().equals(memberId)) {
+            throw new RuntimeException("다른 사람의 블로그입니다.");
+        }
         blogPostRepository.delete(blogPost);
     }
